+++ CHANGELOG.md
@@ -34,6 +34,7 @@
 * Cover `<MatchProfiles>` settings with tests (UIDATIMP-975)
 * Cover `<ProfileAssociator>` component with tests (UIDATIMP-717)
 * Cover `<MappingProfiles/detailsSections/edit>` settings with tests (UIDATIMP-978)
+* Cover `src/utils` utils with tests (UIDATIMP-970)
 
 ## [5.0.1](https://github.com/folio-org/ui-data-import/tree/v5.0.1) (2021-10-19)